requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
google-auth==2.39.0
google-genai==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
packaging==24.2
pgvector==0.2.3
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.19.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-timeout==2.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
requests==2.32.3
rich==13.4.2
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.22
starlette==0.46.2
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
websockets==15.0.1
